Fix off-by-one error in command-line argument handling.
Previously (command-line) did not include the script name, but did
include the executable name if no script was given.  Now if a script
is given its name will be the first element of the list returned by
(command-line) and will be the first element of the list passed to
(main).

This brings us into compliance with SRFI-22.  Our man page was already
correct on this point.